Changes to be made after my first sunum

+++ qa/views.py
@@ -5,6 +5,7 @@ from django.views import View
 from django.core.paginator import Paginator
 from .models import *
 from .forms import *
+from django.db.models import F
 # Create your views here.
 
 
@@ -18,13 +19,19 @@ class QuestionCreateView(LoginRequiredMixin, View):
 
     @staticmethod
     def post(request):
-        form = QuestionForm(request.POST, request.FILES)
+
+        if request.FILES:
+            form = QuestionForm(request.POST, request.FILES)
+        else:
+            form = QuestionForm(request.POST)
+
         if form.is_valid():
             question = form.save(commit=False)
             question.author = request.user
             question.save()
             form.save_m2m()  # saves tags separately
             return redirect('qa:detailed_question', question_id=question.id)
+
         return render(request, 'qa/create_quest.html', {'form': form})
 
 
@@ -75,8 +82,10 @@ class QuestionAllView(View):
 
     @staticmethod
     def get(request):
-        questions = Question.objects.all()
-        paginator = Paginator(questions, 10)
+        # GOT EM ACCORDING TO VOTESTATION
+        # AGGREGATE = 1 TIME ANNOTATE = QUERY SET
+        questions = Question.objects.annotate(vote_score=F('up_votes') - F('down_votes')).order_by('-vote_score')
+        paginator = Paginator(questions, 5)
         page_number = request.GET.get('page')
         questions = paginator.get_page(page_number)
 
@@ -123,12 +132,34 @@ class CommentDeleteView(LoginRequiredMixin, View):
 
     @staticmethod
     def get(request, question_id, comment_id):
+        comment = get_object_or_404(Comment,
+                                    question_id=question_id,
+                                    id=comment_id,
+                                    author=request.user)
+        if request.user != comment.author:
+            return HttpResponseForbidden('You are not authorized to delete this comment')
+        else:
+            try:
+                comment.delete()
+                return JsonResponse({'message': 'Comment deleted'})
+            except Comment.DoesNotExist:
+                return JsonResponse({'message': 'Comment not found'}, status=404)
+            except Exception as e:
+                return JsonResponse({'error': str(e)}, status=500)
+
+
+
+class QuestionDeleteView(LoginRequiredMixin, View):
+    login_url = 'accounts:login'
+
+    @staticmethod
+    def get(request, question_id):
         try:
-            comment = get_object_or_404(Comment, id=comment_id, author=request.user)
-            comment.delete()
-            return JsonResponse({'message': 'Comment deleted'})
-        except Comment.DoesNotExist:
-            return JsonResponse({'message': 'Comment not found'}, status=404)
+            question = get_object_or_404(Question, id=question_id, author=request.user)
+            question.delete()
+            return JsonResponse({'message': 'Question deleted'})
+        except Question.DoesNotExist:
+            return JsonResponse({'message': 'Question not found'}, status=404)
         except Exception as e:
             return JsonResponse({'error': str(e)}, status=500)
